backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
drf-yasg==1.20.0
django-filter==2.4.0
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.9.3
gunicorn==20.0.4
mixer==7.1.2
Pillow==8.3.2
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
django-debug-toolbar==3.2.4
djoser==2.1.0
django-extra-fields